Changing the execution port of the application, setting up the database structure, and creating the authorization logic
Changing the application's execution port from 3333 to 3030 to avoid conflicts. Creating the database structure using the 'knex' library, establishing the user table with the 'id', 'name', 'email', 'password', 'created_at', 'updated_at' fields, and finally the 'role' field of enum type. This field stores information about the type of registered user, indicating whether the user is an 'admin', 'customer', or 'sale'. Lastly, constructing the code responsible for user authorization logic by obtaining the user's 'role' through the 'SessionsController' and adding it to the authentication token. Retrieving this data in the authentication middleware and passing it in the request body, creating the 'verifyUserAuthorization' middleware to check if the user has the necessary 'role' to access a particular functionality, and applying this middleware to the application routes.es. Lines startirng