Updating the front-end by filtering routes and functionalities according to the user type.
Updating the Front-End code. Updating the API address in the 'api.jsx' file. Modifying the application's route structure by creating a route file for each type of user, specifying the available routes for that particular type and rendering according to the user's 'role' attribute. Creating the 'role.jsx' file within the 'utils' folder to standardize user types by creating an object that provides these types to the rest of the application. Finally, filtering available functionalities and pages according to the user's 'role'