Protecting the application data, hiding the user password, and using cookies to store the token
Applying the final updates to the application. Protecting user data by concealing the user password to prevent it from being sent to the Front-End, even when encrypted. Implementing the use of cookies in the application to securely store the user's authentication token, created at the moment when the user logs in. Updating the 'SessionsController.js' file to create the cookie containing the authentication token, and updating the 'server.js' file to configure the 'cors' library to work with cookies. Modifying the 'ensureAuthenticated.js' file to retrieve the authentication token through the cookie passed to the Front-End. Adjusting the 'auth.jsx' file in the Front-End to obtain the authentication token through the cookie sent by the Back-End. Finally, applying the final validations to the application